backend_py/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
pydantic>=2.0.0
httpx>=0.27.0
anthropic>=0.39.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
